feat: :necktie: delete contact

+++ app/contact/controller.js
@@ -47,7 +47,7 @@ update(id2Update, updatedContact) {
       updatedContact,
       {
         returnDocument: "after",
-    runValidators: true,
+        runValidators: true,
       }
     );
   }
@@ -58,8 +58,12 @@ update(id2Update, updatedContact) {
 
 
 delete(id2Delete) {
-  return Contact.findByIdAndDelete(id2Delete);
-},\
+  if (mongoose.Types.ObjectId.isValid(id2Delete)) {
+    return Contact.findByIdAndDelete(id2Delete);
+  }
 
+  // Wrap the error in a rejected promise so that it can be CAUGHT.
+  return Promise.reject(new Error("Invalid ID"));
+},
 
 export default contactController;
